tests/wpt/tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-placeholder-identity.html

<!DOCTYPE html>
<title>shadowrootadoptedstylesheets placeholder identity changes after fetch</title>
<meta name="author" title="Kurt Catti-Schmidt" href="mailto:kschmi@microsoft.com" />
<link rel='help' href='https://github.com/whatwg/html/pull/11687'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='./support/helpers.js'></script>
<body>
<script type="module">
  promise_test(async (t) => {
    // The CSSStyleSheet created before fetch should be the same
    // object from the CSSStyleSheet after fetch completes.
    const url = "./support/styles-red.css";
    const { shadowRoot, testElement } = createStylesheetHost(url);

    assert_equals(shadowRoot.adoptedStyleSheets.length, 1,
      "Before fetch: expected 1 stylesheet(s).");
    assert_equals(shadowRoot.adoptedStyleSheets[0].cssRules.length, 0,
      "Before fetch: stylesheet at index 0 should be empty.");

    const placeholder = shadowRoot.adoptedStyleSheets[0];

    await fetchAndWait(url);

    const fetchedSheet = shadowRoot.adoptedStyleSheets[0];

    assert_equals(fetchedSheet, placeholder,
      "The fetched sheet should not change after fetch.");
    assertSheetRule(shadowRoot, 0, "span { color: red; }", "Fetched sheet");
    assert_equals(placeholder.cssRules.length, 1,
      "The fetched sheet should not be empty.");
    assert_equals(getComputedStyle(testElement).color, "rgb(255, 0, 0)",
      "Red styles from the fetched sheet should be applied.");
  }, "Sheet identity does not change after fetch completes.");
</script>
</body>
